chore(cws): build the ebpfless process resolver on all unix platforms
The macOS Workload Protection collector builds its process tree in memory
from an Endpoint Security message stream, which is exactly what
EBPFLessResolver already does for the ptrace-based ebpfless probe. It has no
/proc or Linux-syscall dependency, so it needs no changes beyond the build
tag.

Widening resolver_ebpfless.go alone is not sufficient, though: on darwin the
process package had zero buildable files, and the resolver depends on two
symbols that lived in Linux-only siblings. Both are pure model.Process field
access with nothing platform-specific about them:

- ResolverOpts and NewResolverOpts move from opts_linux.go to opts_unix.go.
- GetProcessArgv and GetProcessArgv0 move out of resolver_linux.go into a new
  argv_unix.go.

The genuinely Linux-specific helpers in resolver_linux.go (memfd handling,
IsKworker, IsBusybox, setPathname) are not referenced by the ebpfless
resolver and stay where they are.

Note that _unix.go is not a Go filename constraint, so these files carry an
explicit //go:build unix directive; without it they would also build on
Windows and collide with opts_windows.go.

All existing callers of the moved functions are in Linux-only files, and
linux satisfies the unix constraint, so nothing changes for them. BUILD.bazel
regenerated with gazelle.
