+++ pkg/security/resolvers/process/BUILD.bazel
@@ -4,7 +4,8 @@ load("//bazel/rules/go:dd_agent_go_test.bzl", "dd_agent_go_test")
 go_library(
     name = "process",
     srcs = [
-        "opts_linux.go",
+        "argv_unix.go",
+        "opts_unix.go",
         "opts_windows.go",
         "resolver_ebpf.go",
         "resolver_ebpf_test_helpers.go",
@@ -15,6 +16,14 @@ go_library(
     importpath = "github.com/DataDog/datadog-agent/pkg/security/resolvers/process",
     visibility = ["//visibility:public"],
     deps = select({
+        "@rules_go//go/platform:aix": [
+            "//pkg/security/metrics",
+            "//pkg/security/probe/config",
+            "//pkg/security/secl/containerutils",
+            "//pkg/security/secl/model",
+            "//pkg/security/utils",
+            "@com_github_datadog_datadog_go_v5//statsd",
+        ],
         "@rules_go//go/platform:android": [
             "//pkg/discovery/tracermetadata",
             "//pkg/discovery/tracermetadata/model",
@@ -42,6 +51,46 @@ go_library(
             "@com_github_shirou_gopsutil_v4//process",
             "@org_uber_go_atomic//:atomic",
         ],
+        "@rules_go//go/platform:darwin": [
+            "//pkg/security/metrics",
+            "//pkg/security/probe/config",
+            "//pkg/security/secl/containerutils",
+            "//pkg/security/secl/model",
+            "//pkg/security/utils",
+            "@com_github_datadog_datadog_go_v5//statsd",
+        ],
+        "@rules_go//go/platform:dragonfly": [
+            "//pkg/security/metrics",
+            "//pkg/security/probe/config",
+            "//pkg/security/secl/containerutils",
+            "//pkg/security/secl/model",
+            "//pkg/security/utils",
+            "@com_github_datadog_datadog_go_v5//statsd",
+        ],
+        "@rules_go//go/platform:freebsd": [
+            "//pkg/security/metrics",
+            "//pkg/security/probe/config",
+            "//pkg/security/secl/containerutils",
+            "//pkg/security/secl/model",
+            "//pkg/security/utils",
+            "@com_github_datadog_datadog_go_v5//statsd",
+        ],
+        "@rules_go//go/platform:illumos": [
+            "//pkg/security/metrics",
+            "//pkg/security/probe/config",
+            "//pkg/security/secl/containerutils",
+            "//pkg/security/secl/model",
+            "//pkg/security/utils",
+            "@com_github_datadog_datadog_go_v5//statsd",
+        ],
+        "@rules_go//go/platform:ios": [
+            "//pkg/security/metrics",
+            "//pkg/security/probe/config",
+            "//pkg/security/secl/containerutils",
+            "//pkg/security/secl/model",
+            "//pkg/security/utils",
+            "@com_github_datadog_datadog_go_v5//statsd",
+        ],
         "@rules_go//go/platform:linux": [
             "//pkg/discovery/tracermetadata",
             "//pkg/discovery/tracermetadata/model",
@@ -68,6 +117,30 @@ go_library(
             "@com_github_hashicorp_golang_lru_v2//simplelru",
             "@com_github_shirou_gopsutil_v4//process",
             "@org_uber_go_atomic//:atomic",
+        ],
+        "@rules_go//go/platform:netbsd": [
+            "//pkg/security/metrics",
+            "//pkg/security/probe/config",
+            "//pkg/security/secl/containerutils",
+            "//pkg/security/secl/model",
+            "//pkg/security/utils",
+            "@com_github_datadog_datadog_go_v5//statsd",
+        ],
+        "@rules_go//go/platform:openbsd": [
+            "//pkg/security/metrics",
+            "//pkg/security/probe/config",
+            "//pkg/security/secl/containerutils",
+            "//pkg/security/secl/model",
+            "//pkg/security/utils",
+            "@com_github_datadog_datadog_go_v5//statsd",
+        ],
+        "@rules_go//go/platform:solaris": [
+            "//pkg/security/metrics",
+            "//pkg/security/probe/config",
+            "//pkg/security/secl/containerutils",
+            "//pkg/security/secl/model",
+            "//pkg/security/utils",
+            "@com_github_datadog_datadog_go_v5//statsd",
         ],
         "@rules_go//go/platform:windows": [
             "//pkg/process/procutil",
