+++ pkg/security/resolvers/process/opts_unix.go
@@ -3,6 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
+//go:build unix
+
 // Package process holds process related files
 package process
